% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_differential_expression.R
\name{sc_fit_bayesian}
\alias{sc_fit_bayesian}
\title{Fit ZINB Bayesian modelling for Seurat gene expression.}
\usage{
sc_fit_bayesian(
  object,
  feature,
  group.by,
  ctr.ident,
  iter = 1000L,
  warmup = 500L,
  chains = 4L,
  cores = parallel::detectCores(logical = FALSE),
  seed = 42L,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{feature}{Genes to test.}

\item{group.by}{Regroup cells into a different identity class prior to
performing differential expression (see example); "ident" to use Idents}

\item{ctr.ident}{The control identity group.}

\item{iter}{A positive integer specifying the number of iterations for each
chain (including warmup). The default is 1000.}

\item{warmup}{A positive integer specifying the number of warmup (aka
burnin) iterations per chain.}

\item{chains}{A positive integer specifying the number of Markov chains.}

\item{cores}{Number of cores to use when executing the chains in parallel,
defaults to \code{parallel::detectCores}. See: \code{\link[rstan]{sampling}}.}

\item{seed}{The seed for random number generation.}

\item{...}{Additional arguments passed to \code{\link[brms]{brm}}.}
}
\value{
data.frame with ZINB model summary metrics.
}
\description{
Fit ZINB Bayesian modelling for Seurat gene expression.
}
\examples{
\dontrun{
data("pbmc_small")
fit <- (object = pbmc_small, feature = "LYZ", group.by = "groups", ctr.ident = "g1")
}

}
\seealso{
\code{\link[brms]{brm}} for the full list of sampling arguments.
}
