% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratBayesian-package.R
\docType{package}
\name{SeuratBayesian-package}
\alias{SeuratBayesian}
\alias{SeuratBayesian-package}
\title{SeuratBayesian: Bayesian Differential Expression for Seurat via ZINB Models}
\description{
Provides zero-inflated negative bionomial (ZINB) Bayesian modelling for single-cell RNA-seq data stored as Seurat objects.
}
\keyword{internal}
