% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_differential_expression.R
\name{VlnPlot_Bayesian}
\alias{VlnPlot_Bayesian}
\title{Plot Bayesian fit predictions.}
\usage{
VlnPlot_Bayesian(object, feature, group.by, ctr.ident, cols = NULL)
}
\arguments{
\item{object}{A Seurat object.}

\item{feature}{Genes to test.}

\item{group.by}{Regroup cells into a different identity class prior to
performing differential expression (see example); "ident" to use Idents}

\item{ctr.ident}{The control identity group.}

\item{cols}{Colors to use for plotting.}
}
\value{
A ggplot object
}
\description{
Plot Bayesian fit predictions.
}
\examples{
\dontrun{
data("pbmc_small")
VlnPlot_Bayesian(object = pbmc_small, feature = "LYZ", group.by = "groups", ctr.ident = "g1")
}

}
